improve msw mock code + change mock to typescript
