+++ frontend/src/main.jsx
@@ -34,7 +34,7 @@ async function enableMocking() {
     return;
   }
 
-  const { worker } = await import('./mocks/browser');
+  const { worker } = await import('./mocks/browser.ts');
 
   // `worker.start()` returns a Promise that resolves
   // once the Service Worker is up and ready to intercept requests.
