+++ frontend/src/types/systemUser.ts
@@ -3,7 +3,6 @@ import { ServiceResource } from './serviceResource';
 export interface SystemUser {
   id: string;
   integrationTitle: string;
-  description: string;
   systemId: string;
   supplierName: string;
   supplierOrgno: string;
